+++ pkg/query/defaults.go
@@ -13,7 +13,7 @@ var (
 	DefaultFormat         = "txt"
 	DefaultIn             = false
 	DefaultMaxMemPct      = 60
-	DefaultNumResults     = 1000
+	DefaultNumResults     = uint64(1000)
 	DefaultOut            = false
 	DefaultResolveRows    = 25
 	DefaultResolveTimeout = 1 * time.Second
